requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
numpy==1.24.3
